Fix linter errors.
