+++ earningscall/symbols.py
@@ -4,31 +4,12 @@ from collections import defaultdict
 from typing import Optional, Iterator, List
 
 from earningscall.api import get_symbols_v2, is_demo_account
-from earningscall.exchanges import get_exchanges_in_order
+from earningscall.exchanges import exchange_to_index, get_exchanges_in_order, index_to_exchange
 from earningscall.errors import InsufficientApiAccessError
 from earningscall.sectors import sector_to_index, industry_to_index, index_to_sector, index_to_industry
 
-from earningscall.exchanges import FALLBACK_EXCHANGES_IN_ORDER
 
 log = logging.getLogger(__file__)
-
-
-def exchange_to_index(_exchange: Optional[str]) -> int:
-    if not _exchange:
-        return -1
-    try:
-        return get_exchanges_in_order().index(_exchange)
-    except ValueError:
-        return -1
-
-
-def index_to_exchange(_index: int) -> str:
-    if _index == -1:
-        return "UNKNOWN"
-    try:
-        return get_exchanges_in_order()[_index]
-    except IndexError:
-        return "UNKNOWN"
 
 
 security_type_pattern = {
